console/requirements.txt
# AI Media Management Console — Python Dependencies

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Config & validation
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Task queue
celery==5.3.6
redis==5.0.4

# Encryption (OAuth secrets at rest)
cryptography==42.0.5

# HTTP client (for OAuth flows, health checks)
httpx==0.27.0

# Utilities
pyyaml==6.0.1
python-dateutil==2.9.0
